Move `vec_split()` to its own `split.R` file
